refactor(link): promote future API to default entry point
