+++ extensions/link/package.json
@@ -12,16 +12,6 @@
       "types": "./dist/index.d.ts",
       "require": "./dist/index.js",
       "import": "./dist/index.mjs"
-    },
-    "./stable": {
-      "types": "./dist/stable/index.d.ts",
-      "require": "./dist/stable/index.js",
-      "import": "./dist/stable/index.mjs"
-    },
-    "./future": {
-      "types": "./dist/future/index.d.ts",
-      "require": "./dist/future/index.js",
-      "import": "./dist/future/index.mjs"
     }
   },
   "main": "./dist/index.js",
@@ -45,7 +35,6 @@
     "@stackflow/core": "^1.1.1",
     "@stackflow/esbuild-config": "^1.0.3",
     "@stackflow/plugin-history-sync": "^1.7.1",
-    "@stackflow/plugin-preload": "^1.4.3",
     "@stackflow/react": "^1.4.2",
     "@types/react": "^18.3.3",
     "esbuild": "^0.23.0",
@@ -57,7 +46,6 @@
   "peerDependencies": {
     "@stackflow/core": "^1.1.0-canary.0",
     "@stackflow/plugin-history-sync": "^1.6.4-canary.0",
-    "@stackflow/plugin-preload": "^1.4.3-canary.0",
     "@stackflow/react": "^1.3.2-canary.0",
     "@types/react": ">=16.8.0",
     "react": ">=16.8.0"
