--- extensions/link/src/future/index.ts
@@ -1,1 +1,0 @@
-export * from "./Link";
