+++ extensions/link/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./stable";
+export * from "./Link";
