--- extensions/link/src/stable/index.ts
@@ -1,2 +1,0 @@
-export * from "./createLinkComponent";
-export * from "./Link";
